move task result and error into goose.rs
